requirements.txt
python-nmap
beautifulsoup4
